+++ insanic/listeners.py
@@ -5,8 +5,6 @@ import asyncio
 from insanic.conf import settings
 from insanic.connections import _connections
 from insanic.grpc.server import GRPCServer
-from insanic.rabbitmq.connections import RabbitMQConnectionHandler
-from insanic.rabbitmq.helpers import get_callback_function
 from insanic.registration import gateway
 from insanic.services import ServiceRegistry
 
@@ -43,41 +41,11 @@ async def after_server_start_start_grpc(app, loop=None, **kwargs):
         GRPCServer.logger("info", f"GRPC_SERVE is turned off")
 
 
-async def after_server_start_start_rabbitmq_connection(app, loop=None, **kwargs):
-    if settings.RABBITMQ_SERVE:
-        rabbit_mq = RabbitMQConnectionHandler()
-        await rabbit_mq.connect(
-            rabbitmq_username=settings.RABBITMQ_USERNAME,
-            rabbitmq_password=settings.RABBITMQ_PASSWORD,
-            host=settings.RABBITMQ_HOST,
-            port=int(settings.RABBITMQ_PORT),
-            loop=loop
-        )
-        queue_settings = settings.RABBITMQ_QUEUE_SETTINGS
-        for q_s in queue_settings:
-            exchange_name = q_s.get("EXCHANGE_NAME")
-            routing_keys = q_s.get("ROUTING_KEYS", ["#"])
-            queue_name = "_".join([exchange_name] + routing_keys)
-            await rabbit_mq.consume_queue(
-                exchange_name=exchange_name,
-                queue_name=queue_name,
-                routing_keys=routing_keys,
-                callback=get_callback_function(q_s["CALLBACK"]),
-                prefetch_count=q_s.get("PREFETCH_COUNT", 1)
-            )
-    else:
-        RabbitMQConnectionHandler.logger("info", f"RABBITMQ_SERVE is turned off")
-
-
-async def before_server_stop_stop_rabbitmq_connection(app, loop=None, **kwargs):
-    await RabbitMQConnectionHandler.disconnect()
-
-
 async def before_server_stop_stop_grpc(app, loop=None, **kwargs):
     await GRPCServer.stop()
 
 
-def after_server_start_register_service(app, loop, **kwargs):
+async def after_server_start_register_service(app, loop, **kwargs):
     # need to leave session because we need this in hardjwt to get consumer
     gateway.session = aiohttp.ClientSession(connector=aiohttp.TCPConnector(ttl_dns_cache=300))
     gateway.register(app)